BackProp works
Breakthrough. BigDecimal works. Speed is slow. Still have to test if it
gets closer. Might have to leave the program running to see how long it
takes. The csv might be useful. I will merge this branch as soon as the
test will be finished. Hopefully that won’t take long.